+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.2.0
+
 * Added MultiCAS methods for update ops.
 
 ## 0.1.0