+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Added MultiCAS methods for update ops.
 * Added store with fetch and update.
+* Added pubsub for the fields: loading, desc
 
 ## 0.1.0