+++ CHANGELOG.md
@@ -1,8 +1,7 @@
 ## 0.2.0
 
 * Added MultiCAS methods for update ops.
-* Added store with fetch and update.
-* Added pubsub for the fields: loading, desc
+* Added store with fetch and update with pubsub on loading/desc state.
 
 ## 0.1.0